pyproject.toml
[project]
name = "etsy-listing-agent"
version = "0.1.0"
description = "LangGraph-based multi-agent system for Etsy product listing automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "pydantic>=2.0",
    "pandas>=3.0.0",
    "openpyxl>=3.1.5",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "langsmith[anthropic]>=0.6.8",
    "claude-agent-sdk>=0.1.29",
    "google-generativeai>=0.8.6",
    "pillow>=12.1.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/etsy_listing_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.12",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
